+++ packages/sputnik_di/lib/sputnik_di.dart
@@ -3,3 +3,4 @@ library;
 export 'src/lifecycle/lifecycle_export.dart';
 export 'src/state_holder.dart';
 export 'src/deps_node.dart';
+export 'src/dependency.dart';