+++ packages/sputnik_di/lib/src/lifecycle/lifecycle.dart
@@ -17,3 +17,41 @@ abstract class Lifecycle {
   /// Return value: [Future] will complete when all resources are released.
   Future<void> dispose();
 }
+
+/// A basic implementation of the [Lifecycle] interface.
+///
+/// This class allows defining lifecycle behavior through callbacks for initialization
+/// and disposal. It provides a simple way to manage lifecycle events by passing
+/// functions to [onInit] and [onDispose].
+class RawLifecycle implements Lifecycle {
+  /// A function that is called when the lifecycle is initialized.
+  ///
+  /// This function should contain any necessary setup logic.
+  final Future<void> Function() onInit;
+
+  /// A function that is called when the lifecycle is disposed.
+  ///
+  /// This function should contain any cleanup logic.
+  final Future<void> Function() onDispose;
+
+  /// Constructs a [RawLifecycle] with the required lifecycle callbacks.
+  ///
+  /// The [onInit] function is called when [init] is executed, and the [onDispose]
+  /// function is called when [dispose] is executed.
+  RawLifecycle({
+    required this.onInit,
+    required this.onDispose,
+  });
+
+  /// Calls the [onInit] function to initialize the lifecycle.
+  @override
+  Future<void> init() async {
+    await onInit();
+  }
+
+  /// Calls the [onDispose] function to clean up resources.
+  @override
+  Future<void> dispose() async {
+    await onDispose();
+  }
+}