+++ packages/sputnik_di/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.0
+
+- Clearing created dependencies from DepsNode
+- Clearing a specific dependency using the clear() method
+- A new class RawLifecycle for classes that have a lifecycle but do not implement the Lifecycle
+  interface
+- A parameterized singleton factory using bindSingletonFactory
+
 ## 0.0.3
 
 - fixing pub scores